Convert Aggregator to template class
